third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-hr-element-0/hr-expected.txt

This is a testharness.js-based test.
[FAIL] color
  assert_equals: expected "rgb(128, 128, 128)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
